botones arreglados menos guardar

+++ src/components/feed.js
@@ -1,6 +1,6 @@
 /* import { signOut } from 'firebase/auth'; */
 import {
-  post, auth, logOut, addPost, deleteDocData,
+  post, auth, logOut, addPost, deleteDocData, updatePost
 } from '../lib/firebase';
 
 const root = document.getElementById('root');
@@ -67,19 +67,73 @@ export const feed = () => {
       textElement.textContent = feedPosts.text;
       postElement.appendChild(textElement);
 
-      const deleteButton = document.createElement('button');
-      deleteButton.classList.add('delete-btn');
-      deleteButton.textContent = 'Eliminar';
-      deleteButton.value = feedPosts.id;
-      deleteButton.addEventListener('click', () => {
-        const shouldDelete = window.confirm('¿Estás seguro de que deseas eliminar este post?');
-        if (shouldDelete) {
-          deleteDocData(feedPosts.id);
-        }
+      const likeButton = document.createElement('img');
+      likeButton.classList.add('like');
+      likeButton.addEventListener('click', () => {
+        console.log(feedPosts.likes+"like");
       });
-      postElement.appendChild(deleteButton);
+
+      if (feedPosts.userId === auth.currentUser.uid) {
+        const updateButton = document.createElement('img');
+        updateButton.classList.add('update-btn');
+        updateButton.textContent = 'Editar';
+        updateButton.value = feedPosts.id;
+
+        const editSection = document.createElement('section');
+        editSection.classList.add('edit.section');
+        editSection.style.display = 'none';
+        postElement.appendChild(editSection);
+
+        const updateInput = document.createElement('input');
+        updateInput.classList.add('update-input');
+        updateInput.id = feedPosts.id;
+        updateInput.textContent = feedPosts.text;
+        editSection.appendChild(updateInput);
+
+        const saveButton = document.createElement('button');
+        saveButton.classList.add('save-btn');
+        saveButton.textContent = 'Guardar cambio';
+        editSection.appendChild(saveButton);
+
+        const cancelButton = document.createElement('button');
+        saveButton.classList.add('cancel-btn');
+        cancelButton.textContent = 'Cancelar';
+        editSection.appendChild(cancelButton);
+
+        updateButton.addEventListener('click', () => {
+          editSection.style.display = 'block';
+        });
+        saveButton.addEventListener('click', () => {
+          const newPostText = document.getElementById(feedPosts.id);
+          const refPostId = feedPosts.id;
+          console.log(refPostId);
+          updatePost(refPostId, { text: newPostText.value })
+            .then(() => {
+              editSection.style.display = 'none';
+            });
+        });
+        cancelButton.addEventListener('click', () => {
+          editSection.style.display = 'none';
+        });
+        postElement.appendChild(updateButton, editSection);
+
+        // Borrar Post
+        const deleteButton = document.createElement('img');
+        deleteButton.classList.add('delete-btn');
+        deleteButton.textContent = 'Eliminar';
+        deleteButton.value = feedPosts.id;
+        deleteButton.addEventListener('click', () => {
+          const shouldDelete = window.confirm('¿Estás seguro de que deseas eliminar este post?');
+          if (shouldDelete) {
+            deleteDocData(feedPosts.id);
+          }
+        });
+        postElement.appendChild(deleteButton);
+      }
 
       postsContainer.appendChild(postElement);
+      postsContainer.appendChild(postElement);
+      postElement.appendChild(likeButton);
     });
   });
   return feedDiv;
